requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
aspy.yaml==0.3.0
bandit==1.4.0
billiard==3.3.0.23
boltons==16.1.1
brotlipy==0.6.0
cached-property==1.3.0
celery==3.1.25
certifi==2017.4.17
cffi==1.10.0
click==6.7
coverage==4.3.4
decorator==4.0.11
dj-database-url==0.4.2
Django==1.11
django-celery==3.2.1
django-log-request-id==1.3.2
django-model-utils==3.0.0
django-naomi==0.8
django-webpack-loader==0.4.1
django.js-vinta==0.8.2.dev2
fixmydjango==0.2
flake8==2.6.2
flake8-putty==0.4.0
gitdb2==2.0.0
GitPython==2.1.3
gunicorn==19.7.1
ipython==6.0.0
ipython-genutils==0.2.0
isort==4.2.5
jedi==0.10.2
kombu==3.0.37
mccabe==0.5.3
model-mommy==1.3.2
nodeenv==1.1.2
numpy==1.12.1
opbeat==3.5.2
packaging==16.8
pandas==0.19.2
pbr==3.0.0
pexpect==4.2.1
pickleshare==0.7.4
pre-commit==0.13.6
prompt-toolkit==1.0.14
psycopg2==2.7.1
ptyprocess==0.5.1
pycodestyle==2.0.0
pycparser==2.17
pyflakes==1.2.3
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-decouple==3.0
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.13.0
safety==1.4.0
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
smmap2==2.0.1
stevedore==1.21.0
termcolor==1.1.0
traitlets==4.3.2
urllib3==1.21
vine==1.1.3
virtualenv==15.1.0
wcwidth==0.1.7
whitenoise==3.3.0
